Moved public extension. Appending distance to snapshot. matched snapshot filter logic for initial view controller.

+++ surfbreak/surfbreak/Views/Favorite Collection/FavCollectionViewCell.swift
@@ -25,7 +25,7 @@ class FavCollectionViewCell: UICollectionViewCell {
         super.awakeFromNib()
     }
     
-    func setCellContent(waveHeight: Double, waveFrequency: Double, quality: Int, locationName: String, distanceFromUser: Double){
+    func setCellContent(waveHeight: Double, waveFrequency: Double, quality: Int, locationName: String, distanceFromUser: Int){
         heightLabel.text = "\(waveHeight) ft"
         frequencyLabel.text = "\(waveFrequency)s"
         locationLabel.text = locationName
